Finish Chapter 2 Chapter two covers varibales and epressions. This is a pretty basic understanding of assigning variables and performing basic mathemmatical operations.
The basics of Python are quite different from the basics of Javascript for example. I will need to stay aprised that the proetected keywords are different for both languages.

In addition, Python does not have the same use of '^' as in other languages and this requries that we remmeber that upcoming learning will cover XOR.

There was additional information covered with comments, and basics of solving mathematical problems.